script: Move some individual files into subfolders
These are some files that were used in 1 folder. Additionally,
create a new folder for public files. These files are
only used in `lib.rs` and publicly exported. The only
exception is `jsengine`, which was previously in
`script_runtime`, but separated out, since that was also
publicly exported.

Testing: it compiles
