+++ python/tidy/error_message_exceptions.txt
@@ -70,6 +70,7 @@
 ./components/script/dom/indexeddb/idbopendbrequest.rs 1
 ./components/script/dom/indexeddb/idbrequest.rs 7
 ./components/script/dom/indexeddb/idbtransaction.rs 7
+./components/script/dom/indexeddb/key.rs 2
 ./components/script/dom/intersectionobserver/intersectionobserver.rs 2
 ./components/script/dom/macros.rs 2
 ./components/script/dom/navigator/navigator.rs 3
@@ -119,6 +120,5 @@
 ./components/script/dom/xpath/xpathexpression.rs 2
 ./components/script/dom/xpath/xpathresult.rs 1
 ./components/script/drag/drag_data_store.rs 1
-./components/script/indexeddb.rs 2
 ./components/script/xpath.rs 1
 ./components/script_bindings/proxyhandler.rs 1
